feat: D1 이관 Phase 3 — server/src 전체 쿼리 D1 방언 전환 (PLAN_D1_이관.md)
- $N → ?N, now() → SQL_NOW(strftime), ::int 제거, FILTER → SUM(CASE), ILIKE → LIKE + ESCAPE
- json_agg(json_build_object ORDER BY) → 정렬 서브쿼리 + json_group_array, D1 TEXT 반환은 parseJsonCol로 파싱 (API 응답 무변경)
- 의미검색 선행 전환(§7.1/7.2): embedding BLOB(f32 LE) + isolate 벡터 캐시 + JS 코사인, ANY(string_to_array) → json_each
- returned_by_member 0/1 → boolean 매핑, UNIQUE 중복 판별을 메시지 매치로 전환
- items_fts 동기화(§7.3): 등록/수정 DELETE+INSERT batch, 삭제는 items와 동일 batch — group_concat 파생 테이블 별명 필수(스모크에서 발견·수정)
- admin 조회 SELECT에서 embedding BLOB 제외 (ArrayBuffer 직렬화 깨짐, web 미사용)
- scripts 4종은 Neon 전용이라 Phase 5(§8)로 연기

로컬 스모크 전 경로 통과: 탐색·검색·상세, 대여 신청 201/재고 초과 409/반납, 관리자 CRUD·last_admin 가드·allowed_emails. 유닛 44/44 + 전체 타입체크 그린

+++ server/src/services/dashboard.service.ts
@@ -25,14 +25,14 @@ export type DashboardItem = {
 };
 
 export async function getDashboard(db: Sql) {
-  // Q1 전체 물품(폐기 포함) + Q2 현재 대여자 — neon HTTP 드라이버는 쿼리당 1 request라 병렬 실행
+  // Q1 전체 물품(폐기 포함) + Q2 현재 대여자 — D1 도 쿼리당 왕복이라 병렬 실행
   const [itemsRaw, rentersRaw] = await Promise.all([
     db.query(
       `SELECT items.id, items.name, items.kind, items.status, items.total_qty, items.qty_broken,
               (items.total_qty - items.qty_broken) AS rentable_qty,
               (SELECT '/api/photos/' || p.r2_key FROM item_photos p
                 WHERE p.item_id = items.id ORDER BY p.sort_order LIMIT 1) AS photo,
-              (SELECT COALESCE(SUM(r.qty), 0)::int FROM reservations r
+              (SELECT COALESCE(SUM(r.qty), 0) FROM reservations r
                 WHERE r.item_id = items.id AND r.status = 'rented') AS active_now
          FROM items ORDER BY items.id DESC`,
     ),
@@ -45,8 +45,8 @@ export async function getDashboard(db: Sql) {
     ),
   ]);
 
-  // SAFETY: 각 raw 쿼리의 SELECT 목록이 아래 타입과 일치한다. neon HTTP 드라이버의
-  // 반환형이 유니온이라 단언이 필요한데, tsc는 SELECT 문자열을 읽지 못해 이 일치를 검사할 수 없다.
+  // SAFETY: 각 raw 쿼리의 SELECT 목록이 아래 타입과 일치한다. tsc는 SELECT 문자열을 읽지
+  // 못해 이 일치를 검사할 수 없다.
   const items = itemsRaw as unknown as Omit<DashboardItem, "current_renters">[];
   // SAFETY: rentersRaw 도 같은 이유 — Q2 SELECT(member_name·member_phone·qty·item_id)와 타입 일치
   const renters = rentersRaw as unknown as (DashboardRenter & { item_id: number })[];
